requirements.txt
fastapi>=0.100
uvicorn[standard]>=0.22
aiohttp>=3.8
httpx>=0.24
ollama>=0.1.6
prometheus_client>=0.17
psutil>=5.9
pydantic>=1.10
python-dotenv>=1.0
bcrypt==3.2.2
python-jose>=3.3
motor==3.3.1
pymongo==4.5.0
python-dateutil
sendgrid
requests
pydantic[email]
email-validator
PyJWT>=2.8.0
passlib[bcrypt]
psycopg2-binary
mysql-connector-python
pytz